Fix #31 - int is too small for filesize value, use u64
My bad when implementing. Configurator returns u64 and I tried to
put in into an int.

+++ src/DiskFile.h
@@ -75,7 +75,7 @@ protected:
   FILE *handle;
   char *filename;
 
-  void createDiskFile(const std::string &filename, int diskFileSize);
+  void createDiskFile(const std::string &filename, u64 diskFileSize);
   std::string defaultFilename;
 };
 #endif //! defined(__DISKFILE_H__)
